Create README.md
What Makes PTLiga.fans Unique? PTLiga.fans provides an all-encompassing sports experience that goes beyond the basics. The platform covers a wide range of sports, offering live scores, in-depth analyses, and real-time updates. This ensures that fans are always in the know, no matter where they are. The user-friendly interface of PTL